@mssonicbld 2025-05-15
Cherry-pick PR to msft-202412: https://github.com/Azure/sonic-platform-common.msft/pull/84